requirements.txt
scipy==1.4.1
torch==1.6.0
matplotlib==3.1.2
torchvision==0.7.0
numpy==1.18.1
tqdm==4.42.0
pytest==6.1.1
PyYAML==5.3.1